Normalized endian values, version became short, etc.

+++ src/constants.h
@@ -2,9 +2,10 @@
 #define ODICT_CONSTANTS_H_H
 
 #include <string>
+#include "endian.h"
 
 static char ODICT_SIGNATURE[6] = "ODICT";
-static int ODICT_VERSION = 1;
+static short ODICT_VERSION = 1;
 
 static const char* NODE_ETY = "ety";
 static const char* NODE_ENTRY = "entry";
